requirements.txt
beautifulsoup4>=4.12.0
bs4>=0.0.1
certifi>=2023.0.0
chardet>=4.0.0
click>=8.0.1
dateparser>=1.1.8
fake-headers>=1.0.2
greenlet>=2.0.0
html5lib>=1.1
idna>=3.4
lxml>=4.9.0
pathspec==0.9.0
platformdirs==2.3.0
python-dateutil>=2.8.2
pytz>=2023.3
regex>=2023.0.0
requests>=2.31.0
sgmllib3k>=1.0.0
six>=1.16.0
soupsieve>=2.4.0
SQLAlchemy>=2.0.0
tomli==1.2.1
typing-extensions>=4.7.0
tzlocal>=5.0.0
urllib3>=2.0.0
webencodings>=0.5.1
python-dotenv>=1.0.0
